Added support for suspending/restarting comms between nodes
Nodes keep track of an error_count that increments when there is
a network issue while communicating with a peer. If the error count
exceeds a threshold it will stop sending gossip to peer. From there
its up to the peer to send a message that lets the node know to
reset its error_count.