Large rewrite that changes the exchange of peers between nodes.
Nodes can send an initial message that is designed to only
hold the versions of peers the node is aware of. This hopefully
helps limit the amount of data exchanged between nodes. The receiving
node will diff by version before sending full peer data back.

Also, added detection for connection errors. With this, nodes can
track the amount of times connecting to peers failed and use that
to determine some backoff.